+++ tests/dummy/app/app.js
@@ -1,5 +1,5 @@
 import Application from '@ember/application';
-import Resolver from './resolver';
+import Resolver from 'ember-cli-react-fork/resolver';
 import loadInitializers from 'ember-load-initializers';
 import config from './config/environment';